Do not track the spec file, add comments to .gitignore
Since atuin is a rust crate, the 'rust2rpm' tool is being used to
generate an RPM '.spec' file, rather than creating and managing an RPM
spec file by hand. Because of this, any changes that need to be made to
the package should only be made in the rust2rpm.toml file. So, there is
no need to track the spec file, since we can always regenerate it with
rust2rpm.

Some comments were added to the .gitignore file, to make it more clear
why certain files are not being tracked.