Fix url error and add bash-preexec.sh
Not sure if this is a feature with newer versions of rust2rpm, but it
will not allow me to override the url, nor the source-url fields. It
delivers the following message, then quits:

    ERROR: The rust2rpm.toml configuration file specifies an override for
    the package URL, but this is not valid for crates packaged from
    crates.io.

So, those fields were removed, and now rust2rpm must be called like so:

    rust2rpm -V -s atuin $version

The bash-preexec.sh script was also added to this repo
(https://github.com/rcaloras/bash-preexec). According to atuin's docs,
it needs at least this file in order to function
(https://docs.atuin.sh/guide/installation/). Not sure how this spec was
working beforehand (perhaps previous versions of atuin did not require
this?). The file is catted out to %{_buildroot}/etc/profile.d/atuin.sh
as part of the build process, and is only sourced when it detects the
bash shell is in use.

The bash-preexec.sh script is licensed under the MIT license, and so the
license.txt file was not updated, since it already includes that
license. The bash-preexec.sh's exact license was added at the very top
of the file included in this repo.