Refactor how zsh's builtin 'emulate' command is used
This patch removes the second 'emulate [-L ksh]' call in favor of
passing the '-c' flag to run a single line of code outside of ksh.
Directly from the zshbuiltins(1) manpage: "If -c arg appears in flags,
arg is evaluated while the requested emulation is  temporarily  in
effect.  In this case the emulation mode and all options are restored to
their previous values before emulate returns."

The check for the zsh environment was also moved to an elif block,
instead of it's own if statement.